AB-3: Cleanup helpers module and assetsbrowser
