+++ helpers/constants.py
@@ -1,0 +1,11 @@
+"""Helpers Constants"""
+
+SELECTED_FILE = {
+    'File': '',
+    'Path': '',
+}
+FILE_MANAGER = {
+    'Windows': 'Explorer',
+    'Darwin': 'Finder',
+    'Linux': 'File Manager',
+}
